// src/app/interfaces/common/ready-landing-page.interface.ts
export interface ReadyLandingPage {
  _id?: string;
  readOnly?: boolean;
  name?: string;
  videoUrl?: string;
  slug?: string;
  image?: any;
  images?: any;
  productDetails?: any;
  specificationImage?: any;
  whyBestImage?: any;
  certificateImage?: any;
  reviewScreenShoot?: any;
  reviews?: any;
  priority?: number;
  specifications?: any;
  whyBest?: any;
  faqList?: any;
  product?: any;
  description?: string;
  template?: any;
  createdAt?: Date;
  updatedAt?: Date;
  select?: boolean;
}